feat(evaluation): Allow `criteria` to be a callable.
Simplest way I found to support code-based checkpoints.
We can revisit the interface later (thought this looks kind of clean to me).

I added this because I think it would be useful to only rely on Agent/LLM as Judge for cases that can't be handled with simple trace checks.